requirements.txt
pytest==8.4.2
pluggy==1.5.0

numpy==1.26.4        # <— add or change to any <2.0, 1.26.4 is stable
pandas==2.0.0        # keep as-is if you want
scikit-learn>=1.2.0
matplotlib==3.8.4
shap==0.44.1
mlflow>=2.8
pyyaml>=6.0.0
jsonschema>=4.22
Pillow>=10
fastapi>=0.115
uvicorn>=0.30
scipy>=1.10.0




# numpy<2
# pandas
# scipy
# scikit-learn>=1.3
# matplotlib
# shap==0.44.1
# pyyaml
# mlflow>=2.14.0
# evidently
# tabulate
# PyYAML==6.0.2
